backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.3

nltk==3.8.1


tldextract==5.0.1
dnspython==2.4.2
python-whois==0.8.0
validators==0.22.0


email-validator==2.1.0
phonenumbers==8.13.25

Pillow==10.1.0
pytesseract==0.3.10

pydantic==2.5.0
pydantic-settings==2.1.0

httpx==0.25.1
python-jose[cryptography]==3.3.0
cryptography==41.0.7

python-dotenv==1.0.0
aiofiles==23.2.1

pytest==7.4.3
pytest-asyncio==0.21.1